requirements.txt
Django==1.6.2
psycopg2==2.5.2
south==0.8.4
Pillow==2.3.0
yolk3k==0.7
requests
https://github.com/mariocesar/sorl-thumbnail/tarball/master
django-extensions==1.3.3
django-debug-toolbar==1.0.1
django-haystack==2.1.0
pyelasticsearch==0.6.1
djrill
django-suit
paypalrestsdk
django_cron
raven
redis
python-dateutil
django-debug-toolbar-template-timings
pymill
django-haystack-panel
twilio